Split up visitor and visitor tests
